chatbot-backend/requirements.txt
alembic==1.14.0
asyncpg>=0.30.0
fastapi==0.116.1
langchain==0.3.26
langchain-community==0.3.27
langchain-huggingface==0.3.1
langchain-ollama==0.3.2
langchain-postgres==0.0.15
pgvector==0.3.5
psycopg[binary]>=3.0.0
python-dotenv==1.0.0
python-multipart==0.0.20
pypdf==5.8.0
sentence-transformers==5.0.0
sqlalchemy[asyncio]>=2.0.0
uvicorn==0.35.0
